requirements.txt
fastapi
uvicorn
langchain
openai
python-dotenv
docx2txt
pymupdf
firebase-admin
chromadb
